RHealth: A Deep Learning Toolkit for Healthcare Predictive Modeling

RHealth is an open-source R package specifically designed to bring comprehensive deep learning toolkits to the R community for healthcare predictive modeling.

See also

Author

Maintainer: Junyi Gao junyii.gao@gmail.com

Authors:

Ji Song eaad0907@163.com

Zhixia Ren jexica0921@gmail.com

Zhenbang Wu zw12@illinois.edu

John Wu johnwu3@illinois.edu

Chaoqi Yang ycqsjtu@gmail.com

Jimeng Sun jimeng@illinois.edu

Liantao Ma malt@pku.edu.cn

Ewen Harrison ewen.harrison@ed.ac.uk